Created date should feature for duplicated projects next to it
When we create a task we are getting email notification, however when we
click on view task it iis not taking us the task/tool

Also can we please update the term "helper" to a word like a
"Collaborator,"or "Supporting Member," or  "Additional team member" 

Will be good to include the "Estimated hours' as well in the "Home Page"
tab

Under the 'task creation' tab, when we are filling the task details the
date shown in the "estimated start time" tab is random. Can this be
fixed to show that particular's day's date at least, it picks up the
time when it is being opened but the date is always random.

We have a suggetion on replicating the list present under the 'project
category' while we are creating a project, under the tab 'create a task'
as well. This will help in understanding the  time spent on a particular
'task type' under a project during its complete course.